Add throws exceptions in Event and Task services

+++ src/main/java/Calendar/core/services/TaskService.java
@@ -3,12 +3,12 @@ package Calendar.core.services;
 
 import Calendar.core.models.Task;
 import Calendar.core.repositories.TaskRepository;
+import Calendar.exceptions.DataMismatchException;
+import Calendar.exceptions.EntityNotFoundException;
 import Calendar.security.CustomUserDetailsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
 
 @Service
 public class TaskService {
@@ -42,15 +42,17 @@ public class TaskService {
     }
 
     @Transactional
-    public Optional<Task> findById(long id) {
-        return taskRepository.findById(id);
+    public Task findById(long id) {
+        return taskRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Task with id = " + id + " not found"));
     }
 
     @Transactional
     public void deleteById(long id){
-        taskRepository.deleteById(id);
+        if(taskRepository.findById(id).isPresent()) {
+            taskRepository.deleteById(id);
+        }
+        else throw new DataMismatchException("Task with id = " + id + " does not exist");
     }
-
-
 
 }
